backend/requirements.txt
fastapi
uvicorn
supabase
passlib[bcrypt]
python-jose[cryptography]
python-dotenv